Included example dataset

The package includes a dataset resulting of rtweet queries when calling `rtweettree_data()` (to make them available for documentation purposes and offline use).

rtweettree_data_example

Format

An object of class rtweettree_data (inherits from tbl_df, tbl, data.frame) with 12 rows and 93 columns.

Examples

if (FALSE) {
main_status_id <- "1438481824922181635"
rtweettree_data(main_status_id)
}
rtweettree_data_example

#> # A tibble: 12 × 93
#>    type        user_id  status_id  created_at          screen_name text   source
#>  * <chr>       <chr>    <chr>      <dttm>              <chr>       <chr>  <chr> 
#>  1 main_status 1438476… 143848182… 2021-09-16 12:36:07 rtweetbird1 this … Twitt…
#>  2 tree        1438480… 143848428… 2021-09-16 12:45:55 rtweetbird3 this … Twitt…
#>  3 tls         1438480… 143848428… 2021-09-16 12:45:55 rtweetbird3 this … Twitt…
#>  4 tls         1438480… 143848356… 2021-09-16 12:43:01 rtweetbird3 @rtwe… Twitt…
#>  5 tls         1438480… 143848345… 2021-09-16 12:42:36 rtweetbird3 @rtwe… Twitt…
#>  6 like        1438480… 143848428… 2021-09-16 12:45:55 rtweetbird3 this … Twitt…
#>  7 like        1438480… 143848356… 2021-09-16 12:43:01 rtweetbird3 @rtwe… Twitt…
#>  8 like        1438480… 143848345… 2021-09-16 12:42:36 rtweetbird3 @rtwe… Twitt…
#>  9 like        1438479… 143848243… 2021-09-16 12:38:32 rtweetbird2 @rtwe… Twitt…
#> 10 like        1438479… 143848230… 2021-09-16 12:38:02 rtweetbird2 @rtwe… Twitt…
#> 11 like        1438479… 143848243… 2021-09-16 12:38:32 rtweetbird2 @rtwe… Twitt…
#> 12 retweet     1438479… 143848258… 2021-09-16 12:39:09 rtweetbird2 this … Twitt…
#> # … with 86 more variables: display_text_width <dbl>, reply_to_status_id <chr>,
#> #   reply_to_user_id <chr>, reply_to_screen_name <chr>, is_quote <lgl>,
#> #   is_retweet <lgl>, favorite_count <int>, retweet_count <int>,
#> #   quote_count <int>, reply_count <int>, hashtags <list>, symbols <list>,
#> #   urls_url <list>, urls_t.co <list>, urls_expanded_url <list>,
#> #   media_url <list>, media_t.co <list>, media_expanded_url <list>,
#> #   media_type <list>, ext_media_url <list>, ext_media_t.co <list>, …